Split settings by tab
